Refactor Books.Core project structure
