+++ src/books/api/Books.Api/Controllers/BookInstancesController.cs
@@ -1,7 +1,10 @@
-﻿using Books.Core.Queries;
-using Books.Core.Responses;
-using Books.Core.Commands;
+﻿using Books.Core.Create;
+using Books.Core.Delete;
+using Books.Core.GetAll;
+using Books.Core.GetById;
+using Books.Core.GetByISBN;
 using Books.Core.RabitMQ;
+using Books.Core.Update;
 using Books.Infrastructure.Contracts;
 using Books.Infrastructure.Models;
 using MediatR;
