+++ src/books/api/Books.Api/Controllers/BooksController.cs
@@ -1,11 +1,11 @@
-﻿using Books.Core.Commands;
-using Books.Core.Queries;
-using Books.Core.Responses;
-using Books.Core.Commands;
+﻿using Books.Core.Create;
+using Books.Core.Delete;
+using Books.Core.GetAll;
+using Books.Core.GetByISBN;
 using Books.Core.RabitMQ;
+using Books.Core.Update;
 using Books.Infrastructure.Contracts;
 using Books.Infrastructure.Models;
-using Books.Infrastructure.Repositories;
 using MediatR;
 using Microsoft.AspNetCore.Mvc;
