--- src/books/core/Books.Core/Commands/CreateBookInstanceCommand.cs
@@ -1,17 +1,0 @@
-﻿using Books.Core.Responses;
-using Books.Infrastructure.Models;
-using MediatR;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Books.Core.Commands
-{
-    public class CreateBookInstanceCommand : IRequest<string>
-    {
-        public string ISBN { get; set; }
-        public bool IsAvailable { get; set; }
-    }
-}
