+++ src/books/core/Books.Core/Create/CreateBookCommand.cs
@@ -1,14 +1,7 @@
-﻿using Books.Core.Responses;
-using Books.Infrastructure.Models;
+﻿using Books.Infrastructure.Models;
 using MediatR;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using RabbitMQ.Client;
 
-namespace Books.Core.Commands
+namespace Books.Core.Create
 {
     public class CreateBookCommand : IRequest<Book>
     {
