+++ src/books/core/Books.Core/Create/CreateBookHandler.cs
@@ -1,17 +1,9 @@
-﻿using Books.Core.Queries;
-using Books.Core.Responses;
-using Books.Core.Commands;
-using Books.Core.Mapper;
+﻿using Books.Core.Mapper;
 using Books.Infrastructure.Contracts;
 using Books.Infrastructure.Models;
 using MediatR;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace Books.Core.Handlers
+namespace Books.Core.Create
 {
     public class CreateBookHandler : IRequestHandler<CreateBookCommand, Book>
     {
