+++ src/books/core/Books.Core/Create/CreateBookInstanceCommand.cs
@@ -1,0 +1,10 @@
+﻿using MediatR;
+
+namespace Books.Core.Create
+{
+    public class CreateBookInstanceCommand : IRequest<string>
+    {
+        public string ISBN { get; set; }
+        public bool IsAvailable { get; set; }
+    }
+}
