+++ src/books/core/Books.Core/Delete/DeleteBookCommand.cs
@@ -1,0 +1,9 @@
+﻿using MediatR;
+
+namespace Books.Core.Delete
+{
+    public class DeleteBookCommand : IRequest<int>
+    {
+        public string ISBN { get; set; }
+    }
+}
