+++ src/books/core/Books.Core/Delete/DeleteBookInstanceCommand.cs
@@ -1,0 +1,9 @@
+﻿using MediatR;
+
+namespace Books.Core.Delete
+{
+    public class DeleteBookInstanceCommand : IRequest<string>
+    {
+        public int Id { get; set; }
+    }
+}
