+++ src/books/core/Books.Core/GetAll/GetAllBookInstancesQuery.cs
@@ -1,0 +1,10 @@
+﻿using Books.Core.Responses;
+using MediatR;
+
+namespace Books.Core.GetAll
+{
+    public class GetAllBookInstancesQuery : IRequest<List<BookInstanceResponse>>
+    {
+
+    }
+}
