+++ src/books/core/Books.Core/GetAll/GetAllBooksHandler.cs
@@ -1,15 +1,9 @@
-﻿using Books.Core.Queries;
-using Books.Core.Responses;
+﻿using Books.Core.Responses;
 using Books.Infrastructure.Contracts;
 using Books.Infrastructure.Models;
 using MediatR;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace Books.Core.Handlers
+namespace Books.Core.GetAll
 {
     public class GetAllBooksHandler : IRequestHandler<GetAllBooksQuery, List<BookResponse>>
     {
