+++ src/books/core/Books.Core/GetAll/GetAllBooksQuery.cs
@@ -1,0 +1,10 @@
+﻿using Books.Core.Responses;
+using MediatR;
+
+namespace Books.Core.GetAll
+{
+    public class GetAllBooksQuery : IRequest<List<BookResponse>>
+    {
+
+    }
+}
