+++ src/books/core/Books.Core/GetByISBN/GetBookByISBNQuery.cs
@@ -1,0 +1,10 @@
+﻿using Books.Core.Responses;
+using MediatR;
+
+namespace Books.Core.GetByISBN
+{
+    public class GetBookByISBNQuery : IRequest<BookResponse>
+    {
+        public string ISBN { get; set; }
+    }
+}
