+++ src/books/core/Books.Core/GetByISBN/GetBookInstancesByISBNQuery.cs
@@ -1,13 +1,7 @@
 ﻿using Books.Core.Responses;
-using Books.Infrastructure.Models;
 using MediatR;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace Books.Core.Queries
+namespace Books.Core.GetByISBN
 {
     public class GetBookInstancesByISBNQuery : IRequest<List<BookInstanceResponse>>
     {
