+++ src/books/core/Books.Core/GetById/GetBookInstanceByIdQuery.cs
@@ -1,12 +1,7 @@
 ﻿using Books.Core.Responses;
 using MediatR;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace Books.Core.Queries
+namespace Books.Core.GetById
 {
     public class GetBookInstanceByIdQuery : IRequest<BookInstanceResponse>
     {
