--- src/books/core/Books.Core/Handlers/UpdateBookInstanceHandler.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Books.Core.Handlers
-{
-    internal class UpdateBookInstanceHandler
-    {
-    }
-}
