+++ src/books/core/Books.Core/Update/UpdateBookCommand.cs
@@ -1,12 +1,7 @@
 ﻿using Books.Infrastructure.Models;
 using MediatR;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace Books.Core.Commands
+namespace Books.Core.Update
 {
     public class UpdateBookCommand : IRequest<BookRead>
     {
