+++ src/books/tests/Books.CoreTests/Handlers/BooksCommandsHandlersTests.cs
@@ -1,8 +1,9 @@
 ﻿using Xunit;
 using Moq;
-using Books.Core.Commands;
 using Books.Infrastructure.Models;
 using Books.Infrastructure.Contracts;
+using Books.Core.Create;
+using Books.Core.Delete;
 
 namespace Books.Core.Handlers.Tests
 {
